feat: implement JSON data handling in json_handler and integrate with hash_manager
